+++ src/main/java/ch/robinglauser/bfhexercise/game/Updateable.java
@@ -2,5 +2,5 @@ package ch.robinglauser.bfhexercise.game;
 
 public interface Updateable {
 
-    void update(long time);
+    void update(double time);
 }
